feat(intex): pay auction escrow in wCOEN (promis_load strike, u128 amounts)
The auction escrow now locks wrapped COEN. The per-Intex strike becomes the
constant promis_load (the COEN VWAP cancels), replacing the entry-derived strike;
entry_price now feeds only floor/call. IntexAuction._strike is removed and desis
cost_amount_minor() returns promis_load.

wCOEN at 18 decimals pushes escrow amounts past uint64, so lock/refund/paid widen
uint64 -> uint128 across IntexAuction, EscrowAdapter, desis (rate_lock,
ClearingResult) and the REFUND_INSTRUCTIONS wire (abi.encode pads to 32-byte
words, so the wire bytes and golden are unchanged). EscrowAdapter pays wCOEN
(PAYMENT_TOKEN_ALIAS 840 -> 43) and zeroes lockId on a payment-token rotation so
the first deposit re-bootstraps the Compact resource lock. MCP paymentToken -> wCOEN.

Intex settlement (mine-promis) is unchanged: it stays in stables at
entry_price * promis_load (intexfactory derived_cost_amount).

forge test 589 passed; cargo test outbe-desis green.

+++ contracts/intex/src/origin/OriginMessenger.sol
@@ -281,8 +281,8 @@ contract OriginMessenger is
     function quoteSendRefundInstructions(
         uint32 seriesId,
         address[] calldata bidders,
-        uint64[] calldata refundedAmounts,
-        uint64[] calldata paidAmounts,
+        uint128[] calldata refundedAmounts,
+        uint128[] calldata paidAmounts,
         bytes calldata extraOptions,
         bool payInLzToken
     ) external view returns (MessagingFee memory fee) {
@@ -401,8 +401,8 @@ contract OriginMessenger is
     function sendRefundInstructions(
         uint32 seriesId,
         address[] calldata bidders,
-        uint64[] calldata refundedAmounts,
-        uint64[] calldata paidAmounts,
+        uint128[] calldata refundedAmounts,
+        uint128[] calldata paidAmounts,
         bytes calldata extraOptions,
         MessagingFee calldata fee,
         address refundAddress
