+++ contracts/intex/test/foundry/cross-chain/BridgeMsgCodecValidation.t.sol
@@ -220,7 +220,7 @@ contract BridgeMsgCodecValidationTest is Test {
         bytes memory overCap = abi.encodePacked(
             BridgeMsgCodec.BODY_VERSION_V1,
             BridgeMsgCodec.MSG_REFUND_INSTRUCTIONS,
-            abi.encode(uint32(1), new address[](n), new uint64[](n), new uint64[](n))
+            abi.encode(uint32(1), new address[](n), new uint128[](n), new uint128[](n))
         );
         vm.expectRevert(
             abi.encodeWithSelector(
@@ -330,7 +330,7 @@ contract BridgeMsgCodecValidationTest is Test {
     function exposedDecodeRefundInstructions(bytes calldata p)
         external
         pure
-        returns (uint32, address[] memory, uint64[] memory, uint64[] memory)
+        returns (uint32, address[] memory, uint128[] memory, uint128[] memory)
     {
         return BridgeMsgCodec.decodeRefundInstructions(p);
     }
@@ -342,7 +342,7 @@ contract BridgeMsgCodecValidationTest is Test {
     }
 
     function exposedEncodeRefund(uint16 n) external pure returns (bytes memory) {
-        return BridgeMsgCodec.encodeRefundInstructions(1, new address[](n), new uint64[](n), new uint64[](n));
+        return BridgeMsgCodec.encodeRefundInstructions(1, new address[](n), new uint128[](n), new uint128[](n));
     }
 
     function exposedEncodeIssuance(uint16 n) external pure returns (bytes memory) {
