+++ contracts/intex/test/foundry/cross-chain/OriginMessenger.t.sol
@@ -213,9 +213,9 @@ contract OriginMessengerTest is TestHelperOz5 {
     function test_sendRefundInstructions_revert_unauthorized() public {
         address[] memory bidders = new address[](1);
         bidders[0] = address(0x1);
-        uint64[] memory refundedAmounts = new uint64[](1);
+        uint128[] memory refundedAmounts = new uint128[](1);
         refundedAmounts[0] = 100e6;
-        uint64[] memory paidAmounts = new uint64[](1);
+        uint128[] memory paidAmounts = new uint128[](1);
         paidAmounts[0] = 50e6;
 
         MessagingFee memory fee = MessagingFee({nativeFee: 0.1 ether, lzTokenFee: 0});
@@ -272,8 +272,8 @@ contract OriginMessengerTest is TestHelperOz5 {
 
     function test_sendRefundInstructions_revert_empty_array() public {
         address[] memory bidders = new address[](0);
-        uint64[] memory refundedAmounts = new uint64[](0);
-        uint64[] memory paidAmounts = new uint64[](0);
+        uint128[] memory refundedAmounts = new uint128[](0);
+        uint128[] memory paidAmounts = new uint128[](0);
 
         MessagingFee memory fee = MessagingFee({nativeFee: 0.1 ether, lzTokenFee: 0});
         bytes memory options = OptionsBuilder.newOptions().addExecutorLzReceiveOption(200000, 0);
@@ -287,8 +287,8 @@ contract OriginMessengerTest is TestHelperOz5 {
 
     function test_sendRefundInstructions_revert_array_length_mismatch() public {
         address[] memory bidders = new address[](2);
-        uint64[] memory refundedAmounts = new uint64[](2);
-        uint64[] memory paidAmounts = new uint64[](1); // Mismatch
+        uint128[] memory refundedAmounts = new uint128[](2);
+        uint128[] memory paidAmounts = new uint128[](1); // Mismatch
 
         bidders[0] = address(0x1);
         bidders[1] = address(0x2);
@@ -364,8 +364,8 @@ contract OriginMessengerTest is TestHelperOz5 {
 
     function test_quoteSendRefundInstructions() public view {
         address[] memory bidders = new address[](2);
-        uint64[] memory refundedAmounts = new uint64[](2);
-        uint64[] memory paidAmounts = new uint64[](2);
+        uint128[] memory refundedAmounts = new uint128[](2);
+        uint128[] memory paidAmounts = new uint128[](2);
 
         bidders[0] = address(0x1);
         bidders[1] = address(0x2);
