+++ crates/core/desis/src/sol_ext.rs
@@ -92,8 +92,8 @@ sol! {
         function quoteSendRefundInstructions(
             uint32 seriesId,
             address[] calldata bidders,
-            uint64[] calldata refundedAmounts,
-            uint64[] calldata paidAmounts,
+            uint128[] calldata refundedAmounts,
+            uint128[] calldata paidAmounts,
             bytes calldata extraOptions,
             bool payInLzToken
         ) external view returns (MessagingFee memory fee);
@@ -101,8 +101,8 @@ sol! {
         function sendRefundInstructions(
             uint32 seriesId,
             address[] calldata bidders,
-            uint64[] calldata refundedAmounts,
-            uint64[] calldata paidAmounts,
+            uint128[] calldata refundedAmounts,
+            uint128[] calldata paidAmounts,
             bytes calldata extraOptions,
             MessagingFee calldata fee,
             address refundAddress
